#13 Refactored Theta execution code to main
