+++ subprojects/compiler/src/main/kotlin/utils/SystemUtils.kt
@@ -1,0 +1,14 @@
+package hu.bme.mit.semantifyr.oxsts.compiler.utils
+
+import kotlin.reflect.KProperty
+
+open class EnvVar {
+
+    operator fun getValue(thisRef: Any?, property: KProperty<*>): String {
+        return System.getenv(property.name)
+    }
+
+}
+
+@Suppress("ClassName")
+object environment : EnvVar()
